@@ -18,7 +18,7 @@
         }
     },
     "require": {
-        "php": ">=7.3",
+        "php": ">=7.4",
         "hyperf/utils": "^2.1|^3.0",
         "psr/container": "^1.0|^2.0"
     },
