@@ -11,6 +11,7 @@ declare(strict_types=1);
  */
 namespace HyperfTest\Cases;
 
+use Mockery;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -20,6 +21,6 @@ abstract class AbstractTestCase extends TestCase
 {
     protected function tearDown(): void
     {
-        \Mockery::close();
+        Mockery::close();
     }
 }
